fix: refactor a persona (#22)
el delete no se ejecuta por restriccion de FK con empresa, no conozco la logica para saber como proceder

+++ src/main/java/com/coagronet/persona/repositories/PersonaRepository.java
@@ -10,7 +10,7 @@ import com.coagronet.persona.Persona;
 
 public interface PersonaRepository extends JpaRepository<Persona, Long> {
 
-    Page<Persona> findByEstadoIdNot(Integer estadoId, Pageable pageable);
+    Page<Persona> findByEstadoIdNot(Long estadoId, Pageable pageable);
 
     Optional<Persona> findById(Long id);
